🚚(lib-video) move video queries to lib_video
Move video queries to lib_video to be accessible from standalone site.
