+++ src/frontend/packages/lib_video/src/api/useVideos/index.spec.tsx
@@ -1,0 +1,140 @@
+import { renderHook, WrapperComponent } from '@testing-library/react-hooks';
+import fetchMock from 'fetch-mock';
+import { useJwt, videoMockFactory } from 'lib-components';
+import React from 'react';
+import { QueryClient, QueryClientProvider, setLogger } from 'react-query';
+
+import { useVideo, useVideos } from '.';
+
+setLogger({
+  // tslint:disable-next-line:no-console
+  log: console.log,
+  warn: console.warn,
+  // no more errors on the console
+  error: () => {},
+});
+
+jest.mock('lib-components', () => ({
+  ...jest.requireActual('lib-components'),
+  report: jest.fn(),
+}));
+
+let Wrapper: WrapperComponent<Element>;
+
+describe('queries', () => {
+  beforeEach(() => {
+    useJwt.setState({
+      jwt: 'some token',
+    });
+
+    const queryClient = new QueryClient({
+      defaultOptions: {
+        queries: {
+          retry: false,
+        },
+      },
+    });
+
+    Wrapper = ({ children }: Element) => (
+      <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+    );
+  });
+
+  afterEach(() => {
+    fetchMock.restore();
+    jest.resetAllMocks();
+  });
+
+  describe('useVideo', () => {
+    it('requests the resource', async () => {
+      const video = videoMockFactory();
+      fetchMock.mock(`/api/videos/${video.id}/`, video);
+
+      const { result, waitFor } = renderHook(() => useVideo(video.id), {
+        wrapper: Wrapper,
+      });
+      await waitFor(() => result.current.isSuccess);
+
+      expect(fetchMock.lastCall()![0]).toEqual(`/api/videos/${video.id}/`);
+      expect(fetchMock.lastCall()![1]).toEqual({
+        headers: {
+          Authorization: 'Bearer some token',
+          'Content-Type': 'application/json',
+        },
+      });
+      expect(result.current.data).toEqual(video);
+      expect(result.current.status).toEqual('success');
+    });
+
+    it('fails to get the resource', async () => {
+      const video = videoMockFactory();
+      fetchMock.mock(`/api/videos/${video.id}/`, 404);
+
+      const { result, waitFor } = renderHook(() => useVideo(video.id), {
+        wrapper: Wrapper,
+      });
+
+      await waitFor(() => result.current.isError);
+
+      expect(fetchMock.lastCall()![0]).toEqual(`/api/videos/${video.id}/`);
+      expect(fetchMock.lastCall()![1]).toEqual({
+        headers: {
+          Authorization: 'Bearer some token',
+          'Content-Type': 'application/json',
+        },
+      });
+      expect(result.current.data).toEqual(undefined);
+      expect(result.current.status).toEqual('error');
+    });
+  });
+
+  describe('useVideos', () => {
+    it('requests the resource list', async () => {
+      const videos = Array(4).fill(videoMockFactory());
+      fetchMock.mock('/api/videos/?limit=999&organization=1', videos);
+
+      const { result, waitFor } = renderHook(
+        () => useVideos({ organization: '1' }),
+        {
+          wrapper: Wrapper,
+        },
+      );
+      await waitFor(() => result.current.isSuccess);
+
+      expect(fetchMock.lastCall()![0]).toEqual(
+        '/api/videos/?limit=999&organization=1',
+      );
+      expect(fetchMock.lastCall()![1]).toEqual({
+        headers: {
+          Authorization: 'Bearer some token',
+          'Content-Type': 'application/json',
+        },
+      });
+      expect(result.current.data).toEqual(videos);
+      expect(result.current.status).toEqual('success');
+    });
+
+    it('fails to get the resource list', async () => {
+      fetchMock.mock('/api/videos/?limit=999&organization=1', 404);
+
+      const { result, waitFor } = renderHook(
+        () => useVideos({ organization: '1' }),
+        { wrapper: Wrapper },
+      );
+
+      await waitFor(() => result.current.isError);
+
+      expect(fetchMock.lastCall()![0]).toEqual(
+        '/api/videos/?limit=999&organization=1',
+      );
+      expect(fetchMock.lastCall()![1]).toEqual({
+        headers: {
+          Authorization: 'Bearer some token',
+          'Content-Type': 'application/json',
+        },
+      });
+      expect(result.current.data).toEqual(undefined);
+      expect(result.current.status).toEqual('error');
+    });
+  });
+});
